Adicionando validators e consulta pedidos cozinha

+++ src/Controller/pedidos.controller.js
@@ -1,29 +1,29 @@
 import { Pedidos } from "../Model/pedidos.model.js";
-import {
-  atualizarTotalComanda,
-  conferirComandaExecutar,
-} from "./comandas.controller.js";
+import { validationResult } from 'express-validator';
+import { atualizarTotalComanda, conferirComandaExecutar,} from "./comandas.controller.js";
 import { validarMesa } from "./mesas.controller.js";
-import {validarUsuario} from "./usuarios.controller.js"
-import {
-  insert,
-  updateById,
-  deleteById,
-  getById,
-  getAll,
-  dataHora,
-} from "../crud.js";
+import { validarUsuario} from "./usuarios.controller.js"
+import { insert, updateById, deleteById, getById, getAll, dataHora,} from "../crud.js";
 import { inserirItens } from "./itensPedidos.controller.js";
-import { resolveObjectURL } from "buffer";
 
 class PedidosController {
   static async getAllPedidos(req, res) {
     res.json(await getAll(Pedidos));
   }
   static async getPedidoById(req, res) {
-    res.json(await getById(req.body, Pedidos));
+    const errors = validationResult(req)
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() })
+    }
+
+    res.json(await getById(req.params, Pedidos));
   }
   static async postPedido(req, res) {
+    const errors = validationResult(req)
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() })
+    }
+
     let body = req.body;
     body.dataHorario = dataHora();
 
@@ -44,9 +44,24 @@ class PedidosController {
 
   }
   static async putPedido(req, res) {
+    const errors = validationResult(req)
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() })
+    }
+
+    req.body.id = parseInt(req.params.id)
+    
     res.json(await updateById(req.body, Pedidos));
   }
+
   static async deletePedido(req, res) {
+    const errors = validationResult(req)
+    if (!errors.isEmpty()) {
+      return res.status(400).json({ errors: errors.array() })
+    }
+
+    req.body.id = parseInt(req.params.id)
+
     res.json(await deleteById(req.body, Pedidos));
   }
 }
